Query: do not postpone route change when it is older than 10 seconds.
The postponing is needed when a larger bunch of routes comes for one
network one after another. It forces the several route updates to be
processed at once in the end.

On the other hand, it may happen that a network is still being updated
and never gets to be processed. This change fixes this bug.